Refactors for the linter
- Moved `Colour` to separate file to fix circular import
- Comment line length
