+++ apps-rendering/src/palette/background.ts
@@ -18,7 +18,7 @@ import {
 	specialReport,
 	sport,
 } from '@guardian/source-foundations';
-import type { Colour } from '.';
+import type { Colour } from './colour';
 
 // ----- Functions ----- //
 
@@ -33,12 +33,12 @@ const adSlot = (format: ArticleFormat): Colour => {
 	}
 };
 
-const adSlotDark = (_format: ArticleFormat) => neutral[20];
+const adSlotDark = (_format: ArticleFormat): Colour => neutral[20];
 
 const analysisContrastColour = '#f2e8e6';
 const analysisContrastHoverColour = '#e9d9d5';
 
-const mediaArticleBody = (_format: ArticleFormat) => {
+const mediaArticleBody = (_format: ArticleFormat): Colour => {
 	return neutral[10];
 };
