+++ apps-rendering/src/palette/border.ts
@@ -12,7 +12,7 @@ import {
 	specialReport,
 	sport,
 } from '@guardian/source-foundations';
-import type { Colour } from '.';
+import type { Colour } from './colour';
 
 // ----- Functions ----- //
 
@@ -189,21 +189,24 @@ const richLinkSvgDark = (format: ArticleFormat): Colour => {
 };
 
 /**
- *  This is applied server-side. When the page loads, client-side JS applies a class name that overrides this style.
+ * This is applied server-side. When the page loads, client-side JS applies a
+ * class name that overrides this style.
  */
 const richLinkPreload = (_format: ArticleFormat): Colour => {
 	return neutral[60];
 };
 
 /**
- *  This is applied server-side. When the page loads, client-side JS applies a class name that overrides this style.
+ * This is applied server-side. When the page loads, client-side JS applies a
+ * class name that overrides this style.
  */
 const richLinkSvgPreload = (_format: ArticleFormat): Colour => {
 	return neutral[7];
 };
 
 /**
- *  This is applied server-side. When the page loads, client-side JS applies a class name that overrides this style.
+ * This is applied server-side. When the page loads, client-side JS applies a
+ * class name that overrides this style.
  */
 const richLinkSvgPreloadDark = (_format: ArticleFormat): Colour => {
 	return neutral[60];
