+++ apps-rendering/src/palette/colour.ts
@@ -1,0 +1,7 @@
+// ----- Types ----- //
+
+type Colour = string;
+
+// ----- Exports ----- //
+
+export type { Colour };
