+++ apps-rendering/src/palette/hover.ts
@@ -12,7 +12,7 @@ import {
 	specialReport,
 	sport,
 } from '@guardian/source-foundations';
-import type { Colour } from '.';
+import type { Colour } from './colour';
 
 // ----- Functions ----- //
